requirements.txt
mysql55==5.5.38
Django==1.5.1
py26-tastypie @0.9.9_0
python-digest==1.7

djangorestframework==2.3.8
pip install markdown       # Markdown support for the browsable API.
pip install django-filter  # Filtering support

pip install django-rest-swagger